CCIPReader tests cleanup
